backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
boto3==1.34.1
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
neo4j==5.14.1
celery==5.3.4
redis==5.0.1
ollama==0.1.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
ollama==0.1.6
